[playground] rename mongodb local module
Otherwise it gets confusing since we're also using a global `mongodb` module
